Refactored files architecture and commented code
