+++ get_token_list.py
@@ -81,188 +81,3 @@ def get_reverse_index(vocabulary, file):
                     else:
                         reverse_index[token][id_paper] = 1
     return reverse_index
-
-
-def split_query(str):
-    return str.replace("(", "").replace(")", "").split()
-
-
-def split_query_with_p(str):
-    s = str.split()
-    to_delete = 0
-    i = -1
-    l = []
-    for el in s:
-        if "(" in el:
-            to_delete += el.count("(")
-        elif ")" in el:
-            to_delete -= el.count(")")
-            i += el.count(")")
-            if to_delete == 0:
-                l = l + [parentheses_blocks(str)[i]]
-        elif to_delete == 0:
-            l = l + [el]
-    return l
-
-
-def individual_results(index, query):
-    query_list = split_query(query)
-    preresults = {} #Dictionary containing results for each individual term different from AND, OR, NOT
-    for item in query_list:
-        if item not in ["AND", "OR", "NOT"]:
-            if item in index.keys():
-                preresults[item] = index[item].keys()
-            else:
-                preresults[item] = [] #If the word is not in index, we state that no document are relevant for it
-    return preresults
-
-
-def evaluate_single_expression(preresults, single_expr, index):
-    all_docs = docs_in_index(index)
-    for item in single_expr:
-        if item in ["AND", "OR", "NOT"]:
-            if single_expr.index(item) > 0:
-                if single_expr[single_expr.index(item)-1] in ["AND", "OR"]:
-                    previous_word = single_expr[single_expr.index(item) - 2]
-                else:
-                    previous_word = single_expr[single_expr.index(item) - 1]
-                a = preresults[previous_word]
-            if single_expr[single_expr.index(item)+1] == "NOT":
-                next_word = single_expr[single_expr.index(item) + 2]
-            else:
-                next_word = single_expr[single_expr.index(item) + 1]
-            b = preresults[next_word]
-            if item == "AND" and next_word != "NOT":
-                results = intersect(a, b)
-            elif item == "OR" and next_word != "NOT":
-                results = unite(a, b)
-            elif item == "AND" and next_word == "NOT":
-                next_word = single_expr[single_expr.index(item) + 2]
-                b = preresults[next_word]
-                results = remove_in_list(a, b)
-            elif item == "OR" and next_word == "NOT":
-                next_word = single_expr[single_expr.index(item) + 2]
-                b = preresults[next_word]
-                temp = remove_in_list(all_docs, b)
-                results = unite(a, temp)
-            elif item == "NOT" and (single_expr.index(item) == 0 or previous_word not in ["AND", "OR"]):
-                results = remove_in_list(all_docs, b)
-    return results
-
-def evaluate_multiple_expressions(preresults, expr, index):
-    while len(expr) > 1:
-
-        # We build a single expression
-        l = []
-        i = 0
-        while not is_single_exp(l):
-            l = l + [expr[i]]
-            i += 1
-            if i > len(expr):
-                print("Erreur dans la formation d'expressions unitaires")
-                break
-
-        # We evaluate the single expression and get temporary results
-        temporary_results = evaluate_single_expression(preresults, l, index)
-
-        # We now replace the original expression in preresults (which is a dictionary) with a new single key
-        # so we can then work on new terms next to original expression
-        new = l[0]
-        for i in (1, len(l) - 1):
-            if i > 0:
-                new = new + " " + l[i]  # constructing string linked to the expression
-        preresults[new] = temporary_results  # adding said string
-        for item in l:
-            if item not in ["AND", "OR", "NOT"]:
-                preresults = remove_key(preresults, item)  # removing old keys
-
-        # We do the same to the multiple expression list to take into account the evaluation we did in the next loop
-        for item in l:
-            if item not in ["AND", "OR", "NOT"]:
-                expr = remove_in_list(expr, [item])
-        expr = [new] + expr
-        expr = remove_duplicates(expr)
-
-    return temporary_results
-
-
-def parentheses_priority(query):
-    """Recursive function returning the string 'most surrounded'
-    by parentheses which will be treated first later"""
-    if ("(" not in query) and (")" not in query):
-        return query
-    else:
-        s = query[query.find("(") + 1:query.rfind(")")]
-        return parentheses_priority(s)
-
-
-def trimming(query):
-    if "(" not in query:
-        return query
-    else:
-        i = 0
-        while query[i] != "(":
-            i += 1
-        first_pos = i+1
-        i += 1
-        j = 0
-        k = i
-        for c in query[i:]:
-            k += 1
-            if c == "(":
-                j += 1
-            elif c == ")":
-                j -= 1
-                if j == -1:
-                    last_pos = k-1
-                    break
-        return query[first_pos:last_pos], last_pos
-
-
-def parentheses_blocks(query):
-    """Function identifying the different blocks surrounded by parentheses
-    to have a proper order of priority"""
-    blocks = []
-    s = query
-    n = 0
-    if query[0] != "(":
-        blocks = blocks + [query]
-    for c in query:
-        if c == "(":
-            n += 1
-    for i in range(n):
-        new = trimming(s)[0]
-        last_pos_new = trimming(s)[1]
-        blocks = blocks + [new]
-        if trimming(new) == new:
-            s = s[last_pos_new+1:]
-        else:
-            s = new
-    return list(reversed(blocks))
-
-print(parentheses_blocks("(Salut OR (bonjour AND hello) OR (Guten AND Tag))"))
-
-
-def boolean_search(index, query):
-    preresults = individual_results(index, query)
-
-    #We want to treat expressions according to the priority indicated by parentheses
-    blocks = parentheses_blocks(query)
-    for expr in blocks:
-        if "(" not in expr:
-            temp = split_query(expr)
-            temporary_results = evaluate_multiple_expressions(preresults, temp, index)
-            preresults[expr] = temporary_results
-        else:
-            temp = split_query_with_p(expr)
-            temporary_results = evaluate_multiple_expressions(preresults,temp, index)
-
-    final_results = list(set(temporary_results))
-
-    print(str(len(final_results)) + " publications correspondent à votre recherche : ")
-    k = 1
-    for i in final_results:
-        print("\t" + str(k) + "\t Publication n°" + str(i))
-        k += 1
-    return final_results
-
